Remove dead code from kubevirtenv and e2e packages
The PR's back-and-forth left several unused helpers and a near-complete
duplicate of internal/kubevirtenv/binaries.go and docker_kind.go under
test/e2e/. Drop them all and consolidate on the internal package:

- test/e2e: delete kind_cluster.go and the binary catalog / Tools type
  in binaries.go (renamed to repo.go, now only RepoRoot,
  DockerExeFromEnvOrInput, PreparedKindCluster). scenarios_test.go now
  imports DevControllerImageRef from kubevirtenv.
- internal/kubevirtenv/env.go: drop HelmCommand, HelmEnviron,
  ValidateKindInstalled, ValidateClusterctlInstalled,
  capiVersionOrDefault, and the stale DefaultCAPIVersion constant.
  CAPIVersion is now required and set by EnsurePinnedCLIs from the
  pinned clusterctl version; InstallCAPIKubeadmCore errors out
  explicitly when it is empty.
- internal/kubevirtenv/calico.go: drop unused UninstallCalico.
- internal/kubevirtenv/kustomize_apply.go: drop unused DiscardWriter.
- cmd/kubevirt-env/main.go: stop swallowing os.Getwd / FindRepoRoot
  errors in kubevirtEnv(); also drop the redundant CAPIVersion and
  ClusterctlExtraPath assignments.
